fix(go/config): accept fractional image numbers instead of failing the load
The oracle floors whatever JSON number it is given, so timeoutMs 0.5 becomes 1
and artifactsKeepCount 3.9 becomes 3. Decoding those as Go ints did not merely
round differently, it REJECTED them: encoding/json refuses a fractional value
for an int field, which fails the entire config load. A user with one stray
decimal would have found the proxy unable to start.

Both fields are float64 now and floored where the oracle floors. The timeout
floor is raised to 1 rather than allowed to reach zero, because zero means
unset here and a fractional value collapsing to no timeout is not what 0.5 was
asking for. A negative keep count is preserved, since the oracle only requires
the number be finite.

Found by asking what a reviewer would probe next: the oracle takes a JS number
everywhere, so every numeric config field is a place where Go's stricter decode
can turn a shrug into a crash.

+++ go/internal/config/config.go
@@ -104,12 +104,17 @@ type SidecarTimeoutConfig struct {
 // BridgeEnabled defaults to false and is a master switch: the bridge makes
 // PAID upstream image calls, so it is never armed implicitly.
 type ImagesConfig struct {
-	Provider           string `json:"provider,omitempty"`
-	TimeoutMS          int    `json:"timeoutMs,omitempty"`
-	BridgeEnabled      *bool  `json:"bridgeEnabled,omitempty"`
-	BridgeModel        string `json:"bridgeModel,omitempty"`
-	MaxRounds          *int   `json:"maxRounds,omitempty"`
-	ArtifactsKeepCount *int   `json:"artifactsKeepCount,omitempty"`
+	Provider string `json:"provider,omitempty"`
+	// Numeric fields are float64 because the oracle floors whatever JSON
+	// number it is given. Decoding them as int would REJECT a fractional value
+	// outright and fail the whole config load, where the oracle simply floors
+	// it: measured, timeoutMs 0.5 becomes 1 and artifactsKeepCount 3.9
+	// becomes 3.
+	TimeoutMS          float64  `json:"timeoutMs,omitempty"`
+	BridgeEnabled      *bool    `json:"bridgeEnabled,omitempty"`
+	BridgeModel        string   `json:"bridgeModel,omitempty"`
+	MaxRounds          *int     `json:"maxRounds,omitempty"`
+	ArtifactsKeepCount *float64 `json:"artifactsKeepCount,omitempty"`
 }
 
 type VisionSidecarConfig struct {
